Fixed scenario where we only want to update one domain, added option for testing, added support for multiple configs and multiple ip-files

+++ update-ip.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-IP_FILE="/tmp/my-ip"
 
 ## determine where we run
 SOURCE="${BASH_SOURCE[0]}"
@@ -13,13 +12,27 @@ DIR="$( cd -P "$( dirname "$SOURCE" )" >/dev/null 2>&1 && pwd )"
 
 echo "we are in $DIR"
 
-if [ ! -f ${DIR}/setenv ]; then
-    echo "Missing setenv. Aborting"
-    echo "Missing setenv. Aborting" >> ${DIR}/updatelog.txt
+if [ "$1" != "" ]; then
+  if [ -f "$1" ]; then
+    echo "Using config-file $1"
+    echo "Using config-file $1" >> ${DIR}/updatelog.txt
+    CONFIG_FILE=$1
+  else
+    echo "config-file $1 does not exist: falling back to 'setenv'"
+    echo "config-file $1 does not exist: falling back to 'setenv'" >> ${DIR}/updatelog.txt
+    CONFIG_FILE="setenv"
+  fi
+else
+  CONFIG_FILE="setenv"
+fi
+
+if [ ! -f ${DIR}/${CONFIG_FILE} ]; then
+    echo "Missing $CONFIG_FILE. Aborting"
+    echo "Missing $CONFIG_FILE. Aborting" >> ${DIR}/updatelog.txt
     exit 1
  else
     set -e
-    . ${DIR}/setenv
+    . ${DIR}/${CONFIG_FILE}
     set +e
 fi 
 
@@ -29,14 +42,23 @@ if [[ -z $appId || -z $password ]]; then
     exit 1
 fi
 
-az login --service-principal -u $appId --password $password --tenant $tenant
-if [ $? != 0 ]; then
-  echo "an error occurred logging-in. Aborting"
-  exit 1
-fi  
+if [ "$IP_FILE" == "" ]; then
+  IP_FILE="/tmp/my-ip"
+fi
+
+echo "Using IP_FILE=$IP_FILE"
+
+if [ "$TESTING" == "1" ]; then
+  echo "We are simulating. No changes will be made"
+fi
 
 MY_IP=$(curl -s http://whatismijnip.nl |cut -d " " -f 5)
-MY_OLD_IP=$(cat $IP_FILE)
+
+if [ -f "$IP_FILE" ]; then
+  MY_OLD_IP=$(cat $IP_FILE)
+else
+  MY_OLD_IP=""
+fi
 
 if [ "$MY_IP" == "" ]; then
     echo "`date +%F_%R` : IP was empty. Aborting" >> ${DIR}/updatelog.txt
@@ -44,21 +66,42 @@ if [ "$MY_IP" == "" ]; then
 fi
 
 if [ "$MY_IP" != "$MY_OLD_IP" ]; then
+  echo "IP has changed"	
+  az login --service-principal -u $appId --password $password --tenant $tenant
+  if [ $? != 0 ]; then
+    echo "an error occurred logging-in. Aborting"
+    exit 1
+  fi
+
   for REQUESTED_NAME in "${REQUESTED_NAMES[@]}"
   do
+    echo "Processing \"$REQUESTED_NAME\""  
     echo "External IP has changed. '$MY_IP' is not the same as '$MY_OLD_IP'"
-    # more logging    
-    echo "`date +%F_%R` : updating *.${REQUESTED_NAME} and *.${REQUESTED_NAME} with new IP $MY_IP" >> ${DIR}/updatelog.txt
-
-    # delete the entries if there are old IPs
-    az network dns record-set a delete --name *.${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION --yes
-    az network dns record-set a delete --name ${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION --yes
-
-    # update the DNS zone with this IP
-    az network dns record-set a add-record --ipv4-address $MY_IP --record-set-name *.${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
-    az network dns record-set a update --set ttl=60 --name *.${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
-    az network dns record-set a update --set ttl=60 --name ${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
-
+    if [ "$REQUESTED_NAME" != "" ]; then
+      echo "`date +%F_%R` : updating *.${REQUESTED_NAME} and ${REQUESTED_NAME} with new IP $MY_IP" >> ${DIR}/updatelog.txt	    
+      echo "`date +%F_%R` : updating *.${REQUESTED_NAME} and ${REQUESTED_NAME} with new IP $MY_IP"
+      if [ "$TESTING" != "1" ]; then
+        az network dns record-set a delete --name *.${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION --yes
+        az network dns record-set a delete --name ${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION --yes
+        # update the DNS zone with this IP
+        az network dns record-set a add-record --ipv4-address $MY_IP --record-set-name *.${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+        az network dns record-set a add-record --ipv4-address $MY_IP --record-set-name ${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+        az network dns record-set a update --set ttl=60 --name *.${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+        az network dns record-set a update --set ttl=60 --name ${REQUESTED_NAME} --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+      fi	
+    else
+      echo "`date +%F_%R` : updating '*' and '@' with new IP $MY_IP" >> ${DIR}/updatelog.txt
+      echo "`date +%F_%R` : updating '*' and '@' with new IP $MY_IP"
+      if [ "$TESTING" != "1" ]; then
+        az network dns record-set a delete --name "*" --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION --yes
+        az network dns record-set a delete --name "@" --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION --yes
+        # update the DNS zone with this IP
+        az network dns record-set a add-record --ipv4-address $MY_IP --record-set-name "*" --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+        az network dns record-set a add-record --ipv4-address $MY_IP --record-set-name "@" --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+        az network dns record-set a update --set ttl=60 --name "*" --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+        az network dns record-set a update --set ttl=60 --name "@" --resource-group $AZ_DNS_RG --zone-name $PARENT_DOMAIN --subscription $SUBSCRIPTION
+      fi	
+    fi
     echo $MY_IP > $IP_FILE
     echo "`date +%F_%R` : Updated to IP $MY_IP" >> ${DIR}/updatelog.txt
   done  
@@ -68,4 +111,4 @@ else
 	      # more logging    
         echo "`date +%F_%R` : IPs are the same: no action" >> ${DIR}/updatelog.txt
     fi
-
+fi
